requirements.txt
aiodns==2.*,>=2.0.0
aiohttp[speedups]==3.*,>=3.7.0
aiohttp-basicauth==0.*,>=0.2.0
git+https://github.com/TensorTom/aiohttp-csrf@aiohttp3#egg=aiohttp-csrf
aiohttp-debugtoolbar==0.*,>=0.6.0
aiohttp-jinja2==1.*,>=1.4.2
aiohttp-jwt==0.*,>=0.6.1
aiohttp-jwtplus==0.*,>=0.2.0
git+https://github.com/TensorTom/aiohttp-security@master#egg=aiohttp-security[session]
aiohttp-session[aioredis]>=2.9.0
aiojobs==0.*,>=0.3.0
aiologger==0.*,>=0.6.0
aiomysql==0.*,>=0.0.21
git+https://github.com/TensorTom/aioredis.git#egg=aioredis
aioredis-cluster==1.*,>=1.5.2
aiorun==2020.*,>=2020.12.1
apscheduler==3.*,>=3.7.0
beautifulsoup4==4.*,>=4.9.3
cerberus==1.*,>=1.3.2
colorutils==0.*,>=0.3.0
cron-converter==0.*,>=0.2.1
cron-descriptor==1.*,>=1.2.24
cron-schedule-triggers==0.*,>=0.0.11
cron-validator==1.*,>=1.0.3
cronex==0.*,>=0.1.3
croniter==1.*,>=1.0.8
cryptography==3.*,>=3.3.1
demjson==2.*,>=2.2.4
dill==0.*,>=0.3.3
environment==1.*,>=1.0.0
geoip2==4.*,>=4.1.0
gitpython==3.*,>=3.1.14
hemlock-survey==0.*,>=0.0.52
jinja2==2.*,>=2.11.2
jpype1<2.0
js2py==0.*,>=0.70.0
jupyter==1.*,>=1.0.0
llvmlite==0.*,>=0.36.0
mo-dots==4.*,>=4.2.20340
numpy==1.*,>=1.20.1
openquestion==0.*,>=0.1.0
passlib==1.*,>=1.7.4
py4j<1.0
pyjwt==2.*,>=2.0.1
pyprnt==2.*,>=2.5.3
pyqrcodeng==1.*,>=1.3.6
python-arango==7.*,>=7.1.0
python-validators==0.*,>=0.0.1
pyvalidator==0.*,>=0.0.17
pyyaml==5.*,>=5.4.1
quant-survey==0.*,>=0.0.2
questions==0.*,>=0.7.0.a4
redis-wrap==1.*,>=1.4.5
redisgraph==2.*,>=2.3.0
requests==2.*,>=2.0.0
sendgrid==6.*,>=6.5.0
survey==3.*,>=3.4.2
surveyjsunofficial==0.*,>=0.0.1
switch==1.*,>=1.1.0
toml==0.*,>=0.10.2
traversy==0.*,>=0.1.32
git+https://github.com/karls/twilio-python@pyjwt-2.0#egg=twilio
tzcron==1.*,>=1.0.0
ujson==4.*,>=4.0.2
validators==0.*,>=0.18.2
